requirements.txt
telnetlib-313-and-up
pyats[full]
pyyaml
